# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
paths:
  /api/register:
    post:
      summary: 'Register a new user.'
      operationId: registerANewUser
      description: ''
      parameters: []
      responses:
        201:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  user:
                    id: 1
                    name: 'John Doe'
                    email: john.doe@example.com
                    created_at: '2023-07-28 12:00:00'
                    updated_at: '2023-07-28 12:00:00'
                  token: xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: 'John Doe'
                      email:
                        type: string
                        example: john.doe@example.com
                      created_at:
                        type: string
                        example: '2023-07-28 12:00:00'
                      updated_at:
                        type: string
                        example: '2023-07-28 12:00:00'
                  token:
                    type: string
                    example: xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  errors:
                    email:
                      - 'The email has already been taken.'
                properties:
                  errors:
                    type: object
                    properties:
                      email:
                        type: array
                        example:
                          - 'The email has already been taken.'
                        items:
                          type: string
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'The name of the user.'
                  example: 'John Doe'
                email:
                  type: string
                  description: 'The email of the user.'
                  example: john.doe@example.com
                password:
                  type: string
                  description: 'The password of the user.'
                  example: secret
              required:
                - name
                - email
                - password
      security: []
  /api/login:
    post:
      summary: 'Log in a user.'
      operationId: logInAUser
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  user:
                    id: 1
                    name: 'John Doe'
                    email: john.doe@example.com
                    created_at: '2023-07-28 12:00:00'
                    updated_at: '2023-07-28 12:00:00'
                  token: xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: 'John Doe'
                      email:
                        type: string
                        example: john.doe@example.com
                      created_at:
                        type: string
                        example: '2023-07-28 12:00:00'
                      updated_at:
                        type: string
                        example: '2023-07-28 12:00:00'
                  token:
                    type: string
                    example: xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Invalid credentials'
                properties:
                  message:
                    type: string
                    example: 'Invalid credentials'
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: 'The email of the user.'
                  example: john.doe@example.com
                password:
                  type: string
                  description: 'The password of the user.'
                  example: secret
              required:
                - email
                - password
      security: []
...
  /api/problems-and-goals:
    post:
      summary: 'Create Problems and Goals'
      operationId: createProblemsAndGoals
      description: ''
      parameters: []
      responses: {  }
      tags:
        - 'Problems and Goals'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                transcriptId:
                  type: integer
                  description: 'Id of the transcript.'
                  example: 1
              required:
                - transcriptId
      security: []
  '/api/problems-and-goals/{id}':
    post:
      summary: 'Update Problems and Goals'
      operationId: updateProblemsAndGoals
      description: ''
      parameters: []
      responses: {  }
      tags:
        - 'Problems and Goals'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                problemGoalText:
                  type: string
                  description: 'text of the ProblemsAndGoals.'
                  example: rem
              required:
                - problemGoalText
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'Id of the problems and goals.'
        example: 13
        required: true
        schema:
          type: integer
  /api/project-components:
    get:
      summary: 'Get all project components'
      operationId: getAllProjectComponents
      description: 'Get a list of all project components.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      project_id: 1
                      component_id: 1
                      quantity: 3
                      created_at: '2023-07-01T12:00:00Z'
                      updated_at: '2023-07-01T12:30:00Z'
                    -
                      project_id: 1
                      component_id: 2
                      quantity: 2
                      created_at: '2023-07-02T09:00:00Z'
                      updated_at: '2023-07-02T09:30:00Z'
                properties:
                  data:
                    type: array
                    example:
                      -
                        project_id: 1
                        component_id: 1
                        quantity: 3
                        created_at: '2023-07-01T12:00:00Z'
                        updated_at: '2023-07-01T12:30:00Z'
                      -
                        project_id: 1
                        component_id: 2
                        quantity: 2
                        created_at: '2023-07-02T09:00:00Z'
                        updated_at: '2023-07-02T09:30:00Z'
                    items:
                      type: object
                      properties:
                        project_id:
                          type: integer
                          example: 1
                        component_id:
                          type: integer
                          example: 1
                        quantity:
                          type: integer
                          example: 3
                        created_at:
                          type: string
                          example: '2023-07-01T12:00:00Z'
                        updated_at:
                          type: string
                          example: '2023-07-01T12:30:00Z'
      tags:
        - 'Project Components'
      security: []
    post:
      summary: 'Add components to a project'
      operationId: addComponentsToAProject
      description: 'Add website components to a specific project.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      project_id: 1
                      component_id: 1
                      quantity: 3
                      created_at: '2023-07-01T12:00:00Z'
                      updated_at: '2023-07-01T12:30:00Z'
                    -
                      project_id: 1
                      component_id: 2
                      quantity: 2
                      created_at: '2023-07-02T09:00:00Z'
                      updated_at: '2023-07-02T09:30:00Z'
                properties:
                  data:
                    type: array
                    example:
                      -
                        project_id: 1
                        component_id: 1
                        quantity: 3
                        created_at: '2023-07-01T12:00:00Z'
                        updated_at: '2023-07-01T12:30:00Z'
                      -
                        project_id: 1
                        component_id: 2
                        quantity: 2
                        created_at: '2023-07-02T09:00:00Z'
                        updated_at: '2023-07-02T09:30:00Z'
                    items:
                      type: object
                      properties:
                        project_id:
                          type: integer
                          example: 1
                        component_id:
                          type: integer
                          example: 1
                        quantity:
                          type: integer
                          example: 3
                        created_at:
                          type: string
                          example: '2023-07-01T12:00:00Z'
                        updated_at:
                          type: string
                          example: '2023-07-01T12:30:00Z'
      tags:
        - 'Project Components'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                project_id:
                  type: integer
                  description: 'The ID of the project.'
                  example: 1
                components:
                  type: array
                  description: 'An array of component details.'
                  example:
                    - assumenda
                  items:
                    type: string
              required:
                - project_id
                - components
      security: []
...
  /api/project-overview:
    post:
      summary: 'Create Project Overview'
      operationId: createProjectOverview
      description: ''
      parameters: []
      responses: {  }
      tags:
        - 'Project Overview'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                problemGoalID:
                  type: integer
                  description: 'Id of the ProblemsAndGoals.'
                  example: 19
              required:
                - problemGoalID
      security: []
  '/api/project-overview/{id}':
    post:
      summary: 'Update Project Overview'
      operationId: updateProjectOverview
      description: ''
      parameters: []
      responses: {  }
      tags:
        - 'Project Overview'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                overviewText:
                  type: string
                  description: 'text of the ProjectOverview.'
                  example: voluptatem
              required:
                - overviewText
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'Id of the ProjectOverview.'
        example: 19
        required: true
        schema:
          type: integer
  /api/project-summery:
    get:
      summary: 'Get Project Summery List'
      operationId: getProjectSummeryList
      description: ''
      parameters:
        -
          in: query
          name: page
          description: 'page number.'
          example: 3
          required: false
          schema:
            type: integer
            description: 'page number.'
            example: 3
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  current_page: 3
                  data: []
                  first_page_url: 'http://localhost:8000/api/project-summery?page=1'
                  from: null
                  last_page: 1
                  last_page_url: 'http://localhost:8000/api/project-summery?page=1'
                  links:
                    -
                      url: 'http://localhost:8000/api/project-summery?page=2'
                      label: '&laquo; Previous'
                      active: false
                    -
                      url: 'http://localhost:8000/api/project-summery?page=1'
                      label: '1'
                      active: false
                    -
                      url: null
                      label: 'Next &raquo;'
                      active: false
                  next_page_url: null
                  path: 'http://localhost:8000/api/project-summery'
                  per_page: 10
                  prev_page_url: 'http://localhost:8000/api/project-summery?page=2'
                  to: null
                  total: 0
                properties:
                  current_page:
                    type: integer
                    example: 3
                  data:
                    type: array
                    example: []
                  first_page_url:
                    type: string
                    example: 'http://localhost:8000/api/project-summery?page=1'
                  from:
                    type: string
                    example: null
                  last_page:
                    type: integer
                    example: 1
                  last_page_url:
                    type: string
                    example: 'http://localhost:8000/api/project-summery?page=1'
                  links:
                    type: array
                    example:
                      -
                        url: 'http://localhost:8000/api/project-summery?page=2'
                        label: '&laquo; Previous'
                        active: false
                      -
                        url: 'http://localhost:8000/api/project-summery?page=1'
                        label: '1'
                        active: false
                      -
                        url: null
                        label: 'Next &raquo;'
                        active: false
                    items:
                      type: object
                      properties:
                        url:
                          type: string
                          example: 'http://localhost:8000/api/project-summery?page=2'
                        label:
                          type: string
                          example: '&laquo; Previous'
                        active:
                          type: boolean
                          example: false
                  next_page_url:
                    type: string
                    example: null
                  path:
                    type: string
                    example: 'http://localhost:8000/api/project-summery'
                  per_page:
                    type: integer
                    example: 10
                  prev_page_url:
                    type: string
                    example: 'http://localhost:8000/api/project-summery?page=2'
                  to:
                    type: string
                    example: null
                  total:
                    type: integer
                    example: 0
      tags:
        - 'Project Summery'
      security: []
    post:
      summary: 'Create Project Summery'
      operationId: createProjectSummery
      description: ''
      parameters: []
      responses: {  }
      tags:
        - 'Project Summery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                transcriptText:
                  type: string
                  description: 'The text of the transcript.'
                  example: aliquam
                projectName:
                  type: string
                  description: 'The name of the project.'
                  example: non
                clientPhone:
                  type: string
                  description: 'The phone number of the client.'
                  example: eum
                clientEmail:
                  type: string
                  description: 'The email of the client.'
                  example: tosinski@example.org
                clientWebsite:
                  type: string
                  description: 'The website of the client.'
                  example: commodi
              required:
                - transcriptText
                - projectName
      security: []
  '/api/project-summery/{id}':
    get:
      summary: 'Show Project Summery'
      operationId: showProjectSummery
      description: ''
      parameters: []
      responses:
        201:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Data Showed Successfully'
                  data: null
                properties:
                  message:
                    type: string
                    example: 'Data Showed Successfully'
                  data:
                    type: string
                    example: null
      tags:
        - 'Project Summery'
      security: []
    put:
      summary: 'Update Project Summery'
      operationId: updateProjectSummery
      description: ''
      parameters: []
      responses: {  }
      tags:
        - 'Project Summery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                summaryText:
                  type: integer
                  description: 'summaryText of the Project Summery.'
                  example: 14
              required:
                - summaryText
      security: []
    delete:
      summary: 'Delete Project Summery'
      operationId: deleteProjectSummery
      description: ''
      parameters: []
      responses: {  }
      tags:
        - 'Project Summery'
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'Optional parameter. Id of the transcript.'
        required: true
        schema:
          type: integer
        examples:
          omitted:
            summary: 'When the value is omitted'
            value: ''
          present:
            summary: 'When the value is present'
            value: 13
  /api/projects:
    get:
      summary: 'Get all projects'
      operationId: getAllProjects
      description: 'Get a list of all projects.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      project_id: 1
                      project_name: 'Project A'
                      project_description: 'This is project A'
                      total_cost: '1000.00'
                      created_at: '2023-07-01T12:00:00Z'
                      updated_at: '2023-07-01T12:30:00Z'
                    -
                      project_id: 2
                      project_name: 'Project B'
                      project_description: 'This is project B'
                      total_cost: '2000.00'
                      created_at: '2023-07-02T09:00:00Z'
                      updated_at: '2023-07-02T09:30:00Z'
                properties:
                  data:
                    type: array
                    example:
                      -
                        project_id: 1
                        project_name: 'Project A'
                        project_description: 'This is project A'
                        total_cost: '1000.00'
                        created_at: '2023-07-01T12:00:00Z'
                        updated_at: '2023-07-01T12:30:00Z'
                      -
                        project_id: 2
                        project_name: 'Project B'
                        project_description: 'This is project B'
                        total_cost: '2000.00'
                        created_at: '2023-07-02T09:00:00Z'
                        updated_at: '2023-07-02T09:30:00Z'
                    items:
                      type: object
                      properties:
                        project_id:
                          type: integer
                          example: 1
                        project_name:
                          type: string
                          example: 'Project A'
                        project_description:
                          type: string
                          example: 'This is project A'
                        total_cost:
                          type: string
                          example: '1000.00'
                        created_at:
                          type: string
                          example: '2023-07-01T12:00:00Z'
                        updated_at:
                          type: string
                          example: '2023-07-01T12:30:00Z'
      tags:
        - Projects
      security: []
    post:
      summary: 'Store a new project'
      operationId: storeANewProject
      description: 'Create a new project.'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    project_id: 3
                    project_name: 'Project C'
                    project_description: 'This is project C'
                    total_cost: '3000.00'
                    created_at: '2023-07-03T09:00:00Z'
                    updated_at: '2023-07-03T09:30:00Z'
                properties:
                  data:
                    type: object
                    properties:
                      project_id:
                        type: integer
                        example: 3
                      project_name:
                        type: string
                        example: 'Project C'
                      project_description:
                        type: string
                        example: 'This is project C'
                      total_cost:
                        type: string
                        example: '3000.00'
                      created_at:
                        type: string
                        example: '2023-07-03T09:00:00Z'
                      updated_at:
                        type: string
                        example: '2023-07-03T09:30:00Z'
      tags:
        - Projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                project_name:
                  type: string
                  description: 'The name of the project.'
                  example: 'Project C'
                project_description:
                  type: string
                  description: 'The description of the project.'
                  example: 'This is project C'
                total_cost:
                  type: numeric
                  description: 'The total cost of the project.'
                  example: '3000.00'
              required:
                - project_name
                - project_description
                - total_cost
      security: []
...
  /api/scope-of-work:
    post:
      summary: 'Create Scope Of Work'
      operationId: createScopeOfWork
      description: ''
      parameters: []
      responses: {  }
      tags:
        - 'Scope Of Work'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                problemGoalID:
                  type: integer
                  description: 'Id of the ProblemsAndGoals.'
                  example: 8
              required:
                - problemGoalID
      security: []
  '/api/scope-of-work/{id}':
    post:
      summary: 'Update Scope Of Work'
      operationId: updateScopeOfWork
      description: ''
      parameters: []
      responses: {  }
      tags:
        - 'Scope Of Work'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scopeText:
                  type: string
                  description: 'text of the Scope of Work.'
                  example: dolor
              required:
                - scopeText
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'Id of the Scope of Work.'
        example: 11
        required: true
        schema:
          type: integer
...